backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Scraping
scrapy==2.11.0
scrapy-playwright==0.0.33
playwright==1.40.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Task queue and scheduling
celery==5.3.4
redis==5.0.1
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
pandas==2.1.4
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
